+++ Services/SecurityHeadersMiddlewareExtensions.cs
@@ -3,8 +3,8 @@ namespace AgendadorContas.Services;
 public static class SecurityHeadersMiddlewareExtensions
 {
     private const string ContentSecurityPolicy = "default-src 'self'; " +
-        "script-src 'self' 'unsafe-inline'; " +
-        "style-src 'self' 'unsafe-inline'; " +
+        "script-src 'self'; " +
+        "style-src 'self'; " +
         "img-src 'self' data:; " +
         "connect-src 'self'; " +
         "base-uri 'self'; " +